+++ src/main/java/com/example/survey/service/RegisteredUserService.java
@@ -4,6 +4,9 @@ import com.example.survey.dto.RegisteredUserActivationDto;
 import com.example.survey.dto.RegisteredUserChangePasswordDto;
 import com.example.survey.dto.RegisteredUserDto;
 import com.example.survey.dto.RegisteredUserRegistrationDto;
+import com.example.survey.exception.LoginAlreadyInUseException;
+import com.example.survey.exception.MailAlreadyInUseException;
+import com.example.survey.exception.NoSuchRegisteredUserException;
 import com.example.survey.model.RegisteredUser;
 import com.example.survey.model.Role;
 import com.example.survey.repository.RegisteredUserRepository;
@@ -23,7 +26,16 @@ public class RegisteredUserService {
         return registeredUserRepository.findAll();
     }
 
+
     public RegisteredUser addRegisteredUser(RegisteredUserRegistrationDto registeredUserRegistrationDto){
+        Boolean loginExists = registeredUserRepository.existsByLogin(registeredUserRegistrationDto.login());
+        Boolean mailExists = registeredUserRepository.existsByMail(registeredUserRegistrationDto.mail());
+        if(loginExists){
+            throw new LoginAlreadyInUseException("Login "+registeredUserRegistrationDto.login()+" already in use");
+        }
+        if(mailExists){
+            throw new MailAlreadyInUseException("Mail "+registeredUserRegistrationDto.mail()+" already in use");
+        }
         RegisteredUser registeredUser = new RegisteredUser();
 
         Role role = roleRepository.findDistinctByName("registered_user");
@@ -39,13 +51,26 @@ public class RegisteredUserService {
     }
 
 
-    public RegisteredUser getSingleRegisteredUser(Long id) {
+    public RegisteredUser getRegisteredUserById(Long id) {
         return registeredUserRepository.findById(id)
-                .orElseThrow();                             //dopisac obsluge reszty bledow
+                .orElseThrow(() -> new NoSuchRegisteredUserException("No registered user present with id: "+id));                             //dopisac obsluge reszty bledow
     }
 
     public RegisteredUser updateRegisteredUser(RegisteredUserDto registeredUserDto) {
-        RegisteredUser registeredUser = new RegisteredUser();
+        RegisteredUser registeredUser = registeredUserRepository.findById(registeredUserDto.id())
+                .orElseThrow(() -> new NoSuchRegisteredUserException(
+                        "No registered user present with id: "+registeredUserDto.id()
+                ));
+
+        Boolean loginExists = registeredUserRepository.existsByLogin(registeredUserDto.login());
+        Boolean mailExists = registeredUserRepository.existsByMail(registeredUserDto.mail());
+        if(loginExists){
+            throw new LoginAlreadyInUseException("Login "+registeredUserDto.login()+" already in use");
+        }
+        if(mailExists){
+            throw new MailAlreadyInUseException("Mail "+registeredUserDto.mail()+" already in use");
+        }
+
         Role role = roleRepository.findDistinctByName(registeredUserDto.roleName());
 
         registeredUser.setId(registeredUserDto.id());
@@ -56,31 +81,37 @@ public class RegisteredUserService {
         registeredUser.setMail(registeredUserDto.mail());
         registeredUser.setIsActive(registeredUserDto.isActive());
         registeredUser.setRole(role);
-
         return registeredUserRepository.save(registeredUser);
     }
 
     public void deleteRegisteredUser(Long id) {
-        registeredUserRepository.deleteById(id);
+        if(registeredUserRepository.existsById(id))
+            registeredUserRepository.deleteById(id);
     }
 
     public void activateRegisteredUser(RegisteredUserActivationDto registeredUserActivationDto) {
-        RegisteredUser registeredUser = registeredUserRepository.findDistinctById(registeredUserActivationDto.id());
+        RegisteredUser registeredUser = registeredUserRepository.findById(registeredUserActivationDto.id())
+                .orElseThrow(() -> new NoSuchRegisteredUserException(
+                        "No registered user present with id: "+registeredUserActivationDto.id()
+                ));
+
         registeredUser.setIsActive(registeredUserActivationDto.isActive());
-
         registeredUserRepository.save(registeredUser);
-
     }
 
     public void changePassword(RegisteredUserChangePasswordDto registeredUserChangePasswordDto) {
-        RegisteredUser registeredUser = registeredUserRepository.findDistinctById(registeredUserChangePasswordDto.id());
-        registeredUser.setPassword(registeredUserChangePasswordDto.password());
+        RegisteredUser registeredUser = registeredUserRepository
+                .findById(registeredUserChangePasswordDto.id())
+                .orElseThrow(() -> new NoSuchRegisteredUserException(
+                        "No registered user present with id: "+registeredUserChangePasswordDto.id()
+                ));
 
+        registeredUser.setPassword(registeredUserChangePasswordDto.password());
         registeredUserRepository.save(registeredUser);
     }
 
 
-    //getRegisteredUser(): getSingleRegisteredUser() +
+    //getRegisteredUser(): getRegisteredUserById() +
     //updateRegisteredUser() +
     //deleteRegisteredUser() +
     //createRegisteredUser(): addRegisteredUser()+
@@ -90,5 +121,4 @@ public class RegisteredUserService {
     //setCompany()
 
     //todo dodac walidacje pol i obsluge bledow
-    //todo obsluga wyjatkow jako Optionals
 }
